phase 1: 6/8 testing h_sm variant

+++ packboost/cuda/h.cu
@@ -5,10 +5,10 @@
 
 
 // H layout helper: [nfeatsets, nodes, 2, 32]
-static inline __device__ unsigned long long int* Hptr(long long* H, int nodes,
+static inline __device__ uint64_t* Hptr(int64_t* H, int nodes,
                                          int feat, int node, int chan, int lane) {
   size_t idx = (((static_cast<size_t>(feat) * nodes + node) * 2 + chan) * 32u + lane);
-  return (unsigned long long int*)(H + idx);
+  return (uint64_t*)(H + idx);
 }
 
 // ---------------- Kernel (templated on LF dtype) ----------------
@@ -17,7 +17,7 @@ __global__ void _h_sm(
     const uint32_t* __restrict__ XS,  // [nfeatsets, 32*M]
     const int32_t*  __restrict__ Y,   // [N]
     const LF_T*     __restrict__ LF,  // [nfeatsets, N] (u16/u32/u64)
-    long long*      __restrict__ H,   // [nfeatsets, nodes, 2, 32] (int64)
+    int64_t*      __restrict__ H,   // [nfeatsets, nodes, 2, 32] (int64)
     int nfeatsets,
     int cols_32M,     // XS.shape[1] == 32*M
     int N,            // Y.shape[0]
@@ -32,10 +32,10 @@ __global__ void _h_sm(
   const int gwarp      = warps_per_block * blockIdx.y + block_warp;
 
   // Registers for depths 0..2
-  long long hf0=0,  hw0=0;
-  long long hf10=0, hf11=0, hw10=0, hw11=0;
-  long long hf20=0, hf21=0, hf22=0, hf23=0;
-  long long hw20=0, hw21=0, hw22=0, hw23=0;
+  int64_t hf0=0,  hw0=0;
+  int64_t hf10=0, hf11=0, hw10=0, hw11=0;
+  int64_t hf20=0, hf21=0, hf22=0, hf23=0;
+  int64_t hw20=0, hw21=0, hw22=0, hw23=0;
 
   // Shared histogram for depths >=3 : shape [(2^D - 8), 2, 32] int32
   int n_ge3 = (1 << max_depth) - 8;
@@ -83,21 +83,21 @@ __global__ void _h_sm(
           uint32_t       lk = __shfl_sync(mask, l32,    k);
 
           // d = 0
-          hf0 += static_cast<long long>(v) * static_cast<long long>(yk);
+          hf0 += static_cast<int64_t>(v) * static_cast<int64_t>(yk);
           hw0 += v;
 
           // d = 1
           unsigned tk = lk & 1u;
-          if (tk == 0u) { hf10 += v * (long long)yk; hw10 += v; }
-          else          { hf11 += v * (long long)yk; hw11 += v; }
+          if (tk == 0u) { hf10 += v * (int64_t)yk; hw10 += v; }
+          else          { hf11 += v * (int64_t)yk; hw11 += v; }
           lk >>= 1;
 
           // d = 2
           tk = lk & 3u;
-          if      (tk == 0u) { hf20 += v * (long long)yk; hw20 += v; }
-          else if (tk == 1u) { hf21 += v * (long long)yk; hw21 += v; }
-          else if (tk == 2u) { hf22 += v * (long long)yk; hw22 += v; }
-          else               { hf23 += v * (long long)yk; hw23 += v; }
+          if      (tk == 0u) { hf20 += v * (int64_t)yk; hw20 += v; }
+          else if (tk == 1u) { hf21 += v * (int64_t)yk; hw21 += v; }
+          else if (tk == 2u) { hf22 += v * (int64_t)yk; hw22 += v; }
+          else               { hf23 += v * (int64_t)yk; hw23 += v; }
           lk >>= 2;
 
           // d >= 3
@@ -115,22 +115,22 @@ __global__ void _h_sm(
   }
 
   // Write back depths 0..2 (per lane)
-  atomicAdd(Hptr(H, nodes_total, feat_set, 0, 0, lane), (unsigned long long int)hf0);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 0, 1, lane), (unsigned long long int)hw0);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 0, 0, lane), (uint64_t)hf0);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 0, 1, lane), (uint64_t)hw0);
 
-  atomicAdd(Hptr(H, nodes_total, feat_set, 1, 0, lane), (unsigned long long int)hf10);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 2, 0, lane), (unsigned long long int)hf11);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 1, 1, lane), (unsigned long long int)hw10);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 2, 1, lane), (unsigned long long int)hw11);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 1, 0, lane), (uint64_t)hf10);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 2, 0, lane), (uint64_t)hf11);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 1, 1, lane), (uint64_t)hw10);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 2, 1, lane), (uint64_t)hw11);
 
-  atomicAdd(Hptr(H, nodes_total, feat_set, 3, 0, lane), (unsigned long long int)hf20);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 4, 0, lane), (unsigned long long int)hf21);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 5, 0, lane), (unsigned long long int)hf22);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 6, 0, lane), (unsigned long long int)hf23);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 3, 1, lane), (unsigned long long int)hw20);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 4, 1, lane), (unsigned long long int)hw21);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 5, 1, lane), (unsigned long long int)hw22);
-  atomicAdd(Hptr(H, nodes_total, feat_set, 6, 1, lane), (unsigned long long int)hw23);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 3, 0, lane), (uint64_t)hf20);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 4, 0, lane), (uint64_t)hf21);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 5, 0, lane), (uint64_t)hf22);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 6, 0, lane), (uint64_t)hf23);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 3, 1, lane), (uint64_t)hw20);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 4, 1, lane), (uint64_t)hw21);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 5, 1, lane), (uint64_t)hw22);
+  atomicAdd(Hptr(H, nodes_total, feat_set, 6, 1, lane), (uint64_t)hw23);
 
   __syncthreads();
 
@@ -140,10 +140,10 @@ __global__ void _h_sm(
     const int node = 7 + rows_per_warp * block_warp + k;
     if (node < nodes_total) {
       const int base = ((node - 7) * 2) * 32 + lane;
-      const long long fsum = static_cast<long long>(shmem[base + 0]);
-      const long long csum = static_cast<long long>(shmem[base + 32]);
-      atomicAdd(Hptr(H, nodes_total, feat_set, node, 0, lane), (unsigned long long int)fsum);
-      atomicAdd(Hptr(H, nodes_total, feat_set, node, 1, lane), (unsigned long long int)csum);
+      const int64_t fsum = static_cast<int64_t>(shmem[base + 0]);
+      const int64_t csum = static_cast<int64_t>(shmem[base + 32]);
+      atomicAdd(Hptr(H, nodes_total, feat_set, node, 0, lane), (uint64_t)fsum);
+      atomicAdd(Hptr(H, nodes_total, feat_set, node, 1, lane), (uint64_t)csum);
     }
   }
 }
@@ -247,7 +247,7 @@ torch::Tensor h_sm(
       XS_ptr,
       Y.data_ptr<int32_t>(),
       LF.data_ptr<uint16_t>(),
-      H.data_ptr<long long>(),
+      H.data_ptr<int64_t>(),
       nfeatsets, cols_32M, N, max_depth,
       warps_per_block, stride, nodes_tot
     );
@@ -259,7 +259,7 @@ torch::Tensor h_sm(
       XS_ptr,
       Y.data_ptr<int32_t>(),
       LF.data_ptr<uint32_t>(),
-      H.data_ptr<long long>(),
+      H.data_ptr<int64_t>(),
       nfeatsets, cols_32M, N, max_depth,
       warps_per_block, stride, nodes_tot
     );
@@ -271,7 +271,7 @@ torch::Tensor h_sm(
       XS_ptr,
       Y.data_ptr<int32_t>(),
       static_cast<uint64_t*>(LF.data_ptr()),
-      H.data_ptr<long long>(),
+      H.data_ptr<int64_t>(),
       nfeatsets, cols_32M, N, max_depth,
       warps_per_block, stride, nodes_tot
     );
